Move `SavedQueryConfig` to dbt/artifacts
